refactor: scope footer & header translation to components domain
