+++ src/lib/shared/services/i18n/translations/components/footer/en.ts
